moved AttestationParticipants to rightful location
